% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_meta.R
\name{get_sed_meta}
\alias{get_sed_meta}
\title{Get sepectral data from .sed file}
\usage{
get_sed_meta(path)
}
\arguments{
\item{path}{The path where the .sed file is located}
}
\value{
A data frame with column names equal to wavenumbers and one row with reflectance values
}
\description{
Get sepectral data from .sed file
}
